Expose version at runtime
This commit allows programs that depend on `purescript` to ask what
version of it they've got at runtime. This was motivated by Pursuit,
where we want to show which version of a particular package is being
used. Specifically, we want to be able to say things like "this is the
documentation for the prelude in PureScript 0.6.8".

refs #purescript/pursuit/34