Provide more information in `exportedDctors`
Currently, `exportedDctors` returns only a [ProperName] with the names
of the data constructors. This is no good if you want to know about
the arguments of each constructor. This commit fixes the issue by
changing the return type to [(ProperName, [Type])].

On a related note, I wonder if the documentation for `isExported` and
`isDctorExported` should encourage you to use `exportedDecls` and
`exportedDctors`, as these are 'higher level' versions of what is really
the same function, which won't require you to pattern match (and
therefore hopefully be safer and less likely to break because of
updates).