desugar imports in psc-docs
psc-docs now requires all modules to be passed on the command line so
that it can performs the desugarImports step of desugaring. This is
necessary for it to be able to work out which instances should be
displayed in the output.

Accordingly, a new --docgen option has been added, similar to psc's
--codegen, which allows you to specify the modules you want
documentation for. See psc-docs --help for more information.

This commit also removes the 'Module Documentation' header.