Collect fixity information in docs
Collect fixity information in the first documentation rendering pass,
and also use it in the output of psc-docs.

Operators with no fixity declarations are given the default fixity,
which is currently infixl, -1 precedence. (notably, it is not possible
to get this precedence with an explicit fixity declaration; this should
possibly be revisited later)