Add psc-publish warning for unacceptable versions
A problem arose with a package that depended on purescript-maybe, and
was using the version 0.3.0-rc.1. Haskell's Data.Version refuses to
parse this as a Version, which means purescript-maybe was missing from
the resolvedDependencies.

Now, if the resolved Bower version for dependencies is not parseable
into a Version, a warning will be emitted, although it will still be
missing from resolvedDependencies.

In the longer term, we should look at using a Version data type that
supports semver properly, which would probably eliminate this issue
completely.