Add the compiler version to psc-publish output
For pursuit: to determine whether a given JSON-serialized package was
produced by a sufficiently recent compiler, so that it can fail
gracefully if necessary.

We also modify the parsers to accept an extra argument which is the
minimum allowable version. The FromJSON instance sets the minimum
allowable version to 0.0.0.0 (that is, it never fails with a
CompilerTooOld error).