Move the bulk of bundling code into the library
This moves the bulk of psc-bundle into the purescript library, so that
projects besides purescript can take advantage of this code without having
to execute the psc-bundle executable.

Most of the code from psc-bundle/Main.hs was moved directly to
src/Language/PureScript/Bundle.hs.  The @app@ function was split
apart, with the parsing and generation portion put into the library
while the loading from disk was left in Main.hs.