adds a loadAll command to psc-ide
This command loads every compiled module it detects inside the output
folder. This command is somewhat of a necessity to make the import
lookup useful.

We can now start the server, load all the compiled modules and have
import completion for all the modules in our project. Previously we'd
have to either load a file that already imported a certain module or we
had to scan the output folder ourself.