Improve license errors in psc-publish
Currently, you get the 'no license specified' error even if you do
specify a license but it's not a valid SPDX license expression. There
is quite a lot of code in the wild which specifies a license but does
not use the SPDX format, and in these cases the error is not very good.

This commit adds a new error type to differentiate between these two
cases. It also removes the (now redundant) LicenseNotFound constructor.