Safer installation from source in INSTALL.md
This commit ensures that the stackage snapshot used when building from
source following the instructions in INSTALL.md will be the same as the
one used in development and in binary bundles.

This is safer (because you'll end up with a set of dependencies which
have been tested together) and it also means you'll always get the
latest version according to Hackage. In comparison, the Stackage nightly
is often behind because of waiting for purescript or other libraries to
make changes to allow all packages in the nightly to build together.