Add "--format html" to psc-docs
Following on from #2520. This commit adds a --format html option to
psc-docs which produces HTML documentation in a given directory.

All HTML pages are self-contained (in particular, the CSS is duplicated
in every page; it's not very large though, so I don't think this is a
problem), and all links are relative, so it should be easy to just stick
these HTML files pretty much anywhere and have them 'just work'.

The CSS is copied from Pursuit's CSS, with a minor change made via
inline styles where we want to differ slightly from what Pursuit does in
the header.

We also generate an 'index.html' file with a list of all the modules in
the documentation bundle.