Attempt to address CI timeout issues
* Separate installing GHC and compiling dependencies from the build
  step, so that we can apply a timeout to the former.
* Use the `timeout` command so that builds fail if they take too long.

Hopefully this should address CI timeout issues: if dependencies change
e.g. after switching compilers and it takes too long to compile them
all, we make the build fail with a bit of time to spare so that the
cache can be uploaded, and then the build can be manually restarted, so
that the next attempt can continue where the previous one left off.

This solution is based on an observation that Travis uploads the build
cache on failed builds, provided that the build failed for some other
reason than it being timed out (if a build times out, it doesn't get the
opportunity to upload the build cache).

This commit re-enables the OSX Travis builds since hopefully they will
work now.

This commit also removes the hack of setting the `compiler` variable
in the build matrix so that each job in the build matrix gets a separate
cache; this is no longer necessary as Travis now maintains separate
caches for separate jobs.