Type operator desugaring now more precise.
As per feedback, the desugaring of type operators has been localised
to type class instance declarations only.

There is a test now to ensure that the children will appear for both
the type that has the aliased operator as well as the type class
itself.

Additionally there appeared to be a bug in the testing framework with
regard to the `ShouldBeDocumented` constructor. It was performing an
intersection of the actual and expected lists and this was always passing
for all inputs. At least in my testing. Changing it to a straight equality
test gives what I expect to be the correct behaviour.

Sorry for rolling that in with this change but it seemed related and
appropriate.

This is an example of how the docs appear now. The `Boop` type has an
alias operator of (</>) that is used in the type class instance declaration.

http://jsbin.com/hopogohowi