Fix child decl ordering (#2984)
* Fix child declarations order in HTML docs

Child declarations - that is, data constructors, type class members, and
instances - were appearing in the reverse order because of how the
`partitionChildren` function works.

This fixes https://github.com/purescript/pursuit/issues/125.

* Test that order of child declarations is preserved

This doesn't fully test what we want it to, as it is still possible to
accidentally reverse the list during HTML rendering. However that's not
too difficult to verify manually (as I have just done).