[purs ide] Adds an "editor-mode"
When the editor-mode flag is specified at startup the server will not start a
file watcher process anymore. Instead it only reloads after succesful rebuild
commands. This is a lot less fragile than relying on the file system APIs, but
will mean that a manual load needs to be triggered after builds that didn't go
through purs ide.

* Also updates the README under psc-ide/ to be a bit more recent