Don't use `Show` instance for pretty printing
The `Show` instance was not following the intent of the type class.
https://hackage.haskell.org/package/base-4.10.0.0/docs/Text-Show.html#t:Show

We weren't making a "syntactically correct Haskell expression."
Instead, we were using it to pretty print the representation we wanted.

Rather than conflate the ideas,
we use a function specific for pretty printing.
The name might not be the best.
But, it more correctly reflects the intent/purpose.