Updates from review
Firstly, there's a little change to the error output in the case of
there being no place for the instance except Prim. Previously, it just
printed an empty list, which didn't make much sense, gramatically. Now,
however, a different error is displayed to point out that there's no
reasonable home for the instance.

As well as this, the `typeModule` function in `TypeChecker.hs` can now
deal with proxy types. Previously, it couldn't, and the tests were
passing really due to an accident of laziness. Now, the tests pass, and
all is well in the world.