Add valid location list to orphan instance error (#3106)
* Add valid location list to orphan instance error

Previously, the orphan instance simply said to consider a newtype. Now,
it lists all the places in the project that would be considered valid
locations for the instance as well. This hopefully closes #3063.

* Updates from review

Firstly, there's a little change to the error output in the case of
there being no place for the instance except Prim. Previously, it just
printed an empty list, which didn't make much sense, gramatically. Now,
however, a different error is displayed to point out that there's no
reasonable home for the instance.

As well as this, the `typeModule` function in `TypeChecker.hs` can now
deal with proxy types. Previously, it couldn't, and the tests were
passing really due to an accident of laziness. Now, the tests pass, and
all is well in the world.

* Tidy the orphan instance error further.

Thinking about it, you'll only ever see, at most, two things in the list
of possible locations. So, we can intercalate the list with `or` and get
a much friendlier-looking error.

* Adjust wording on error feedback

Previously, the error mentioned that the instance "must" be declared in
one of a set of places, which @goodacre.liam (rightly) noted would imply
that there aren't other, equally valid, solutions. The wording has been
altered to address this.