Tag let bindings with metadata about their origin
When we parsed where clauses, we desugared directly to a let binding.
When we did that desugaring,
we lost the information about whether a let binding was ever a where clause.

We add a bit of metadata to the let bindings we generate
so we can track whether it began its life as a where clause or not.

This is already showing to be directly useful,
as we can pretty print where clauses!
Should allow external tools to also benefit from the metadata.