Implement `Coercible` for safe zero-cost coercions
* Define a new built-in module, `Prim.Coerce` which exports a single
type class, `Coercible a b`, which relates types with identical runtime
representations.

* Extend the type checker's class solver to solve constraints of the
form `Coercible a b` automatically, as per the rules in the paper "Safe
Zero-Cost Coercions for Haskell" (though simpler due to the absence of
some of Haskell's features like type families).

* As part of the above, introduce the notion of roles to the compiler. Roles
will be inferred for all PureScript types but can also be declared explicitly.