Fix handling of Prim submodules with docs
Fixes #3347

This commit mainly consists of re-export handling so `purs docs` doesn't
choke on modules which re-export things from submodules of Prim.

It also fixes an issue with the builtin Prim docs which meant that the
RowToList class was being considered a type, leading to incorrect
information in the rendered HTML docs, as well as internal errors in
certain situations.

Finally, we fix handling of links to Prim submodules, ensuring that all
builtin modules are actually considered builtin, not just Prim. This has
no effect on `purs docs` but it will affect Pursuit, once we get around
to updating it.

This commit also adds a test which ensures that generating docs for a
module which re-exports things from Prim works.