Use `microlens-platform` instead of `lens`
`lens` is pretty big of a dependency with many transitive dependencies.
Build times with `lens` are kind of long.

We use a small subset of `lens`,
almost everything we use also exists in `microlens`.
The few things we don't use are available in
`microlens-ghc`, `microlens-mtl`, and `microlens-th`.
And `microlens-platform` bundles all of that stuff together.
`microlens-platform` is a much smaller dependency.

We _could_ drop down to `microlens` (which purports a 3.5s build time).
but that would take some bigger changes.
This change was small and quick,
and that seems ideal to see if it's a worthwhile change.