Fix check for valid identifiers during JS codegen
Fixes #3481; the check for whether a string represents a valid JS
identifier was previously too permissive, i.e. it would allow through
some strings are not valid as JS identifiers, leading to invalid syntax
in the generated JS. For example, a string beginning with a digit is
not a valid JS identifier but it was previously considered to be by
the codegen phase of the compiler.