Fix isUnquotedKey check to allow reserved names (#3487)
Fixes #3482. We have allowed reserved names (like `data` or `type`) to
be used as record labels without being quoted since #699 was resolved.
This commit makes the `isUnquotedKey` function aware of this, having the
effect that these labels are no longer unnecessarily quoted in error
messages. For example, if we enter

    > {} :: { data :: Int }

in the repl, we now receive:

    Error found:

      Type of expression lacks required label data.

    while checking that expression {}
      has type { data :: Int
               }

whereas previously, `data` would have been quoted.