Simplify docs re-export handling (#3534)
Follows on from #3521 by removing the use of the Env's imports and
instead obtaining the same information from the Env's exports. By
removing the need for the Imports values, this paves the way for using
externs files to speed up docs generation and make it more resilient
(see #3503). Additionally, it should offer a small performance boost
immediately.

This code can probably still be simplified and optimised a bit more. For
example, we could use a more sensible data structure for storing
Declaration values as they are constructured, in preparation for later
lookups while we are adding re-exports.