Truncate types in errors, fixes #3401
This commit truncates types in error messages below a depth of 3 levels
in the type AST. If the --verbose-errors flag is provided, the
truncation depth is instead set to 1000 (the same as with
pretty-printing of values). Suggestions and docs code is not affected,
since in both of these cases, we always want to print the entire type.

In the process, I've done a minor refactoring of the PrettyPrintType
type to change how rows and records are represented, to make them a
little easier to deal with.