Simplify `purs publish` resolutions format
Fixes #3499.

Simplifies the format for resolutions files passed via the CLI to `purs
publish`. A new-style resolutions file should look something like this:

```
{
  "purescript-prelude": {
     "version": "4.0.0",
     "path": "bower_components/purescript-prelude"
  },
  "purescript-lists": {
     "version": "6.0.0",
     "path": "bower_components/purescript-lists"
  },
  ...
}
```

The version field is used for generating links between packages on
Pursuit, and the path field is used to obtain the source files while
generating documentation: all files matching the glob "src/**/*.purs"
relative to the `path` directory will be picked up.

The "version" field is optional, but omitting it will mean that no links
will be generated for any declarations from that package on Pursuit. The
"path" field is required.

We continue to permit the old-style format (i.e. the output of
`bower list --json`), but eventually we will want to start warning when
we encounter it, and eventually drop support for it entirely.

Some other smaller changes which were made in the process of
implementing the above:

- Use a new directory `tests/purs/publish` for packages for testing
`purs publish`, as opposed to using packages from the
`tests/support/bower_components` directory. This is necessary if we want
to be able to test packages which have any dependencies (previously we
were only testing with prelude, which has none). The new directory
contains copies of the prelude and console libraries with dependencies
pre-installed.

- Stop emitting warnings about extraneous or undeclared dependencies; we
no longer have enough information to be able to do this.