Consider constraints when selecting instances
Previously, instance selection ignored instance constraints when
determining if instances overlapped, and considered all instances with
matching type arguments. This meant that `instance x :: SomeClass X` and
`instance y :: (SomeConstraint a) => SomeClass a` were considered
overlapping even if `X` doesn't satisfy `SomeConstraint`.

This commit relaxes the check on instance declarations to permit
instances to overlap in a limited way: for each type in the instance, a
concrete type is allowed to overlap with a type variable if the type
variable appears in an instance constraint. Two type variables, even if
given different constraints, are not allowed to overlap in any case.

This commit also modifies instance selection to attempt to solve for all
the overlapping instances with matching type arguments, and if exactly
one instance can be solved, select it instead of throwing an overlapping
instance error. If zero or more than one have solutions, the original
overlapping instance error is thrown.

The upshot is that there are now some (more?) cases where truly
ambiguous overlapping instances are detected at use instead of at
declaration, but in exchange, some cases of nonambiguous "overlapping"
instances, such as the above example, are now permitted.