Fail compilation when a module imports itself (#3586)
Fixes #3079, fixes #3118. This is a breaking change so we should wait
until after 0.12.4 is released to merge.

I have used the CycleInModules error to indicate this; previously we
only emitted this error for cycles containing two or more modules, but
now we also emit it for singleton cycles (which occur precisely when
modules import themselves).

This commit also fixes a separate issue where the CycleInModules error
was not printing one of the modules in a cycle (specifically, the module
which happened to come first in the cycle), presumably due to
accidentally using the tail rather than the whole list while
constructing the error message. I've made this problem easier to avoid
here by using `nonEmpty` instead of `:|` to construct the non-empty
list, which means that we don't need to bind a reference to the tail at
all.