Use usual Unicode format for illegal whitespace errors
For example, if you have a non-breaking space in a source file:

Before

    Unable to parse module:
    Illegal whitespace character '\160'

After

    Unable to parse module:
    Illegal whitespace character U+00A0

The latter is a much more common format for presenting this information,
and therefore should be e.g. easier to google.

Text.Printf is in base now so using it shouldn't lead to any dependency
difficulties.