Improve parser error ErrRecordUpdateInCtr
When parsing a record literal, if we encounter a = after a label, we
emit a parse error, because = is only for record updates.

However, the error says "Expected ':', saw '='", which implies that ':'
would be the only valid token following a record label, although
a label could also be followed by '}' in a record pun.