Use externs files when producing docs
Resolves #3503, fixes #3624.

This commit does a few related things:

**Docs codegen target**

Adds a `docs` codegen target for `purs compile`. When supplied, a
`docs.json` file is written for each module alongside the other files in
that module's directory within the output directory. The file contains a
JSON-serialized Docs.Types.Module, which is the same as the one which
will be produced by `purs publish` for that module, except that the
re-exports field will be empty; this is because computing re-exports
requires access to documentation for all of the upstream modules, and we
don't have an easy way of getting hold of all of these (which won't
adversely affect memory consumption / performance for the regular
compilation path).

Since we are producing docs during `purs compile`, we now have easy
access to the relevant externs files, which we can use to fill in type
signatures of any declarations which don't already have their own. This
means that no module within the `Language.PureScript.Docs` namespace
needs to call into the type checker any more in order to get accurate
and useful type signatures in docs, regardless of whether annotations
have been given or not, which both speeds up docs generation
significantly and eradicates the possibility of bugs like #3624 arising.

**Use docs.json files in purs docs and purs publish**

The subcommands `purs docs` and `purs publish` now each accept a
`--compile-output` option, which should point to a compiler output
directory, with default value `"output"`. With this information, we can
reach into the compiler output directory to grab the information in each
of the required `docs.json` files, perform one final pass over them to
fill in re-exports, and then produce documentation in whatever format is
requested. If the `docs.json` files are not up-to-date with respect to
the source files which have been provided to `purs docs`, we produce
them (just as if we had run `purs compile --codegen docs`).

In effect, `purs docs` and `purs publish` now behave a little more like
`purs bundle` does, in the sense of getting their input data from the
output of `purs compile`, rather than from a collection of source files.

**Refactor Docs library API**

The Docs library API has been refactored and simplified; now, the `purs
docs` CLI and the `Language.PureScript.Publish` module only really need
to use one function for getting hold of docs: the function `collectDocs`
from the module `Language.PureScript.Docs.Collect`. This function
implements the behaviour described above for reading `docs.json` files,
making sure they are up-to-date, and filling in re-exports.

With this new `collectDocs` function, a number of other functions no
longer need to be exposed as part of the compiler library API. In
particular, the module `Language.PureScript.Docs.ParseInPackage` has
been removed entirely.