Add failing test for conflicting re-exports (#3502)
Within a module,
we cannot define a class and a data type with the same name.
The issue here is that we can define a class and a data type with
the same name in different modules,
then re-export them both from one module.

We don't want to allow this behavior.
We add a failing test to codify that this should not work.