Disallow re-exporting same class/type name (#3502)
As mentioned in the previous commit
(09bfea484ff6e0d77c028f11314763c794df8c47),
we don't want to allow re-exporting a class and a type that have
the same name.
The changes here make that a reality.

The majority of the tests are breaking because we're dependent on
`purescript-typelevel-prelude` and it violates the changes here.
See: https://github.com/purescript/purescript-typelevel-prelude/issues/43
for more information.