Print the type of the conflicting name (#3502)
Before, we only had conflicting exports for the same types of things.
Now that we disallow exports of different types,
we'll want to be a bit more explict in what the problem is.

Errors we'd get before looked like:
```
Export for type B.X conflicts with A.X
```
While that does convey what the problem is, we can do a bit better.

Errors now look like:
```
Export for type B.X conflicts with type class A.X
```