Fix multiple "let"s in ado before the final "in"
This failed because we were checking the layout stack before collapsing
it. Subsequent let statements were tagged as a normal LytLet context
instead of a LytLetStmt context because the previous LytLetStmt was on
top of the stack instead of LytAdo/LytDo. By collapsing first, it pops
the previous LytLetStmt, and we get the LytAdo/LytDo on top again.