Update to GHC 8.6.5, lts 13.26
I'm hoping this will fix an access violation seen in Windows CI:

    <https://travis-ci.org/purescript/purescript/jobs/549447718>

but updating things is probably a good idea in general anyway.

I've tightened the lower bound on language-javascript to 0.6.0.13,
because version 0.6.0.12 (the version in the snapshot we're now using)
has a bug where it does not consider `as` to be a valid identifier.