Add a placeholder purs.bin file to the repo
This is necessary for successful publishing, because the npm cli will
attempt to read this file once before the postinstall script is run (at
which point it will be replaced with the actual binary).

I've also made this file produce a sensible error if you try to execute
it, which wasn't previously the case (the current one is just a text
file).