Refactor and simplify BuildPlan a little
Instead of having two separate MVars for build job results and errors,
just have one, which contains a sum type, to indicate if and how a build
job has completed with a little more clarity and safety (in the sense
that this makes some invalid states unrepresentable).

Additionally, rather than having two separate functions for consuming
the result of a build plan, namely `collectErrors` and `collectResults`,
and requiring that the first is called before the second, unify them
both into `collectResults`.

This will help for #3145, as for that I need the BuildPlan to be able to
expose which build jobs succeeded before their errors are rethrown, so
that we can store their timestamps and hashes in preparation for the
next build.

I haven't done any performance tests on this just yet, but I don't
anticipate any drastic changes.