`errorSuggestion` allows for multiple suggestions
Previously errorSuggestion returned `(Just "")` for UnusedImport and
DuplicateImport and `Nothing` for any other ErrorMessage that has no
suggestions. Both `Just ""` and `Nothing` seem to be treated the same
and represented as `""` in the rendered versions.

The matching imports suggestions for UnknownName now are represented as
seperate suggestions, which allow for proper rendering.

The JSON represenation of multiple suggestions is still a single string,
as changing that would require changing the JSON schema.