Move `SimpleErrorMessage` to `Errors`
In 4828b5d3fa078361e03dd0f3dbbfa46dd4ee6b20, we moved
`Language.PureScript.Errors.SimpleErrorMessage` to
`Language.PureScript.AST.Declarations`. Per the comment in the Pull
Request:
https://github.com/purescript/purescript/pull/2230#issuecomment-233145477,
we did this because we wanted to move
`Language.PureScript.Errors.ErrorMessageHint` into
`Language.PureScript.AST.Declarations.Expr`.

Thing is, we didn't have to move everything just to move
`Language.PureScript.Errors.ErrorMessageHint`. In fact, by making this
move, we forced the AST to require knowledge about every type of error
that can be created. For example, the AST has to know about parse
errors (something that should happen before creating an AST) and code
generation errors (something that should happen after creating an AST).
In an ideal world, the AST would only know about errors that come about
from modifications to the AST itself.

The real issue that this led to was that the AST had to know about the
errors that came about from the CST. Given that the CST is the surface
syntax of the AST, that requirement seems a bit off. With the
dependencies inverted this way, it's hard to decouple the AST from the
CST.

We take a step towards the ideal world by moving
`Language.PureScript.AST.Declarations.SimpleErrorMessage` back into
`Language.PureScript.Errors`. This should allow us to extract the AST
from the `purescript` package itself such that it can be built upon in
isolation from the rest of the package.

We should revisit why
`Language.PureScript.AST.Declarations.SimpleErrorMessage` is so massive
in the future. It will probably make sense to separate the data type
into the different categories of errors so those errors can live closer
to where they should be constructed.