Split the constants based on Prim and Prelude
Prim stuff is part of PureScript the language, Prelude is part of the
PureScript conventions. PureScript the language doesn't require an
existence of Prelude in order to work properly. Instead of mixing these
two levels, we separate the constants so we can pull the Prim constants
out with the AST.

In order to not break everything, we re-export the constants from the
original module. We should put a timer on this and remove the re-exports
after that point so we don't wind up with a bunch of cruft like the
other re-export modules.