Extract AST to its own package
We've gone quite a ways with a sinle package for everything PureScript
related. This has led to incredibly long compile times if you want to
consume PureScript the library, as you have to compile everything no
matter how much you want to really use. To help with downstream
consumers, we start splitting into packages of useful functionality.

We start with the AST as this _should_ be the very basis of everything
PureScript related. From the AST, someone ought to be able to build
packages atop it that add more functionality to the PureScript language.
For instance, we can split out the CST as a package that builds atop the
AST and adds more functionality (namely, parsing).

It's maybe a little disheartening that there don't appear to be any
tests around the AST. Maybe with things separated out a bit more, it'll
be more inviting to write some tests.

The modules we move here might not be all that intuitive, but they seem
to be the core of what the PureScript AST is all about. The only
questionable modules are the `Control.Monad.Supply` and
`Control.Monad.Supply.Class` modules. We could flip the dependency in
`Language.PureScript.Names` and move those values into the
`Control.Monad.Supply.Class` module. It's unclear which way to go.

Since the `Control.Monad.Supply.Class` module isn't in the
`Language.PureScript` hierarchy, it might make sense for it to stay in
the top-level. However, it might be too general for that and should
ideally live in its own package (maybe even outside the project?).

We also use PackageImports judiciously. When trying to figure out which
dependencies went where, it was much harder than it should have been to
figure this stuff out. When trying to figure out which dependencies were
used and necessary, it was even harder to figure what the situation was.
Finally, when trying to figure out where `Control.Monad.Supply.Class`
came from, it took a frustratingly long amount of time to realize it was
a module we had defined in the project.

For all of these reasons, we use PackageImports to help future us know
which dependencies are used, and where. Of course there are tools like
weeder (https://github.com/ndmitchell/weeder) to help with finding
unused dependencies and whatnot. But, it's also good to know statically
what is being used and to allow other tools to be used.

Finally, we update the license-generator to work with this split off
package. We might have to do more work before we're done, but we can at
least generate the license file in a similar manner to how we do at the
top-level.