Extract the CST to a package
We extract out the CST to its own package that builds atop the AST. This
should make it a bit easier to build other tooling atop the CST. In
particular, anything that might need to parse PureScript files can now
depend on this package to do so.

Since this new package only depends on `purescript-ast` (and its
transitive dependencies), it can be built in a reasonable amount of
time.

It's important to note that we don't pull out `Language.PureScript.CST`.
That module is straddling the line between a re-exporting "prelude" and
an adapter between `purescript-ast` and `purescript`. Since it deals
with errors, and errors currently exist at the top-level, we cannot yet
move this module down into `purescript-cst`.

Probably what needs to happen is that we make a split between
`Language.PureScript.CST` being a re-exporter and
`Language.PureScript.CST` doing adapter work. Not clear which way to go
on this one as both seem viable avenues.

It's really great to see a test suite around some of this deep down
functionality. Maybe we can use this example to feed back into the AST
and make a test suite there.