Setup sdists for each package
Now that we have more than one package, we have to do things a little
different. `stack sdist` will create an sdist for each package. This
means we end up with a directory that looks like:

```console
$ tree sdist-test
sdist-test
├── purescript-0.13.6.tar.gz
├── purescript-ast-0.1.0.0.tar.gz
└── purescript-cst-0.1.0.0.tar.gz

0 directories, 3 files
```

If we try to use `tar -xzf sdist-test/purescript-*.tar.gz -C sdist-test
--strip-components=1`, the glob will expand to all three sdists and fail
with a cryptic error:

```console
$ tar -xzf sdist-test/purescript-0.13.6.tar.gz sdist-test/purescript-ast-0.1.0.0.tar.gz sdist-test/purescript-cst-0.1.0.0.tar.gz -C sdist-test --strip-components=1
tar: sdist-test/purescript-ast-0.1.0.0.tar.gz: Not found in archive
tar: sdist-test/purescript-cst-0.1.0.0.tar.gz: Not found in archive
tar: Exiting with failure status due to previous errors
```

We change to mirroring the directory structure within the `sdist-test`
directory, and running that way.

It's a bit unfortunate that the top-level package has to be handled
differently, but this should help us move forward.