Fix DuplicateModule.purs golden test
System.FilePath.Glob explicitly states that the results of a glob are
not in any defined order. However, the order of the files provided to
L.P.Make.make determines which file is reported as having a duplicate
module. So in order to make the output of DuplicateModule.purs
deterministic (and possibly other error messages as well), the test
framework sorts the input files passed to make when testing the text of
warnings and errors.