Setup `hlint`
There have been a smattering of efforts to run `hlint` on the project:
* 1b77d7e83bac4c83e57358cf6b5d8f2f9cbe7f6f
* afd4605dcbfe3710a8b9252b07a583fce8f18f62
* 5860f1ee38cbdc3edb5cabd75bd22fc894807a76
* b120af74c689e491911909532172c5b315b5f86e
* a647c7893e64b4b5886e5fe9d8c95a468aece820
* f806a6f4d51c8917c754269ad382327bd161febe
* 20f1200532678510ff2ec4ee28d376619c51e3c8
* 0b07beb08d68b3ae1c77d645f85ef681cc260b9d

The last explicit `hlint` inspiried cleanup was three and a half years
ago. Over that time, we've accumulated almost 400 hints from `hlint`.
It's unclear if these are all new or if they've been around for a while
and the previous efforts only addressed some of the issues.

To mitigate the number of new hints, we setup `hlint` properly in the
project. Given that there are too many hints to attempt to take on all
at once, we ignore all of the current hints with a `.hlint.yaml` file
provided by the `--default` flag. We can triage the hints that we think
are useful, and ignore the ones we don't find useful.

The motivation behind doing this change now is that we want to be more
consistent with the language extensions in use:
https://github.com/purescript/purescript/issues/3805. Since `hlint` can
enforce the use of language extensions:
https://github.com/ndmitchell/hlint/tree/9f1e23655c4fac537446a2600dd2f5d89825406c#restricting-items,
it seemed like an alright approach to solving this problem:
https://github.com/purescript/purescript/issues/3805#issuecomment-599297787.

We'll move on to cleaning up language extensions after we've decided
what to do about this approach.